Veneer-migrate 3 structured/failure dispatch methods (batch 17)
Phase 2 batch 17 for Issue #682. Migrates 3 more impl Agent dispatch
methods: the structured verifier selection pair + the failure-path
handler.

Moved (turn.rs impl Agent -> verifier_orchestration.rs free fn):

  - handle_structured_task_contract_verifier_selection (62 LOC) —
    materializes Python package markers, runs the structured
    verifier inside a Spinner scope, dispatches to the finish helper
    on success or the transport-error-to-outcome path on Err.
  - finish_structured_task_contract_verifier_selection (56 LOC) —
    emits agent.autotest.completed, detects + masks external import
    contamination, builds auto_test feedback, records the verifier
    invocation, seeds the artifact ledger with the Pass/Fail
    observation, and observes the bound exit-zero on success.
  - handle_task_contract_verifier_failure (89 LOC) — increments the
    verifier-failure retry counter, optionally promotes the repair
    job (or emits safe-stop on budget exhaust), classifies rerun
    progress, and queues the system-note retry prompt.

Adjustments:

  * verifier_orchestration.rs:
    - Added super::auto_test::AutoTestResult to the existing
      super::auto_test::{AutoTestPlan, VerifierCommand} import block
      (the other ~12 deps — verifier_repair_context_from_failure,
      apply_verifier_rerun_observation, VerifierOutcome,
      structured_verifier_invocation_report,
      detect_verifier_external_import_contamination,
      run_structured_task_contract_verifier,
      task_contract_verifier_transport_error_to_outcome,
      task_contract_verifier_outcome_label,
      task_contract_auto_test_result_to_outcome,
      build_feedback_for_auto_test,
      Spinner, etc — are already imported from earlier batches or
      reachable via super::*).

  * turn.rs:
    - Promoted 5 Agent methods to pub(super) so the relocated
      veneers can call them via agent.method(...):
      materialize_python_package_markers_for_owned_test_imports,
      materialize_python_package_markers_for_external_import,
      emit_agent_verifier_invoked_if_new,
      emit_agent_verifier_external_import_rejected_if_first,
      observe_task_contract_verifier_exit_zero_bound.
    - Deleted the 3 original impl Agent method bodies (207 LOC).
    - Dropped now-unused turn.rs imports from
      super::verifier_driver:: (detect_verifier_external_import_contamination,
      run_structured_task_contract_verifier,
      structured_verifier_invocation_report,
      task_contract_auto_test_result_to_outcome,
      task_contract_verifier_outcome_label,
      task_contract_verifier_transport_error_to_outcome) and
      StructuredTaskContractVerifierRun from
      super::verifier_orchestration:: re-import (only the moved fns
      referenced them).
    - Re-added task_contract_verifier_transport_error_to_outcome
      under crate::agent::loop_run::verifier_driver:: inside
      mod tests for the
      verifier_timeout_transport_error_becomes_verifier_failure_evidence
      regression test fixture (2 call sites preserved).

Metrics: turn.rs 32,510 -> 32,295 (-215);
verifier_orchestration.rs 2,585 -> 2,809 (+224). Cumulative Issue
#682 progress: turn.rs 34,958 -> 32,295 (-2,663, ~89% of Phase 2
target). Acceptance gap: 32,295 -> 32,000 = ~295 LOC remaining.
